+++ marketback/marketgod/src/main/java/com/th/marketgod/mapper/ReservationMapper.java
@@ -20,5 +20,5 @@ import org.apache.ibatis.annotations.Param;
 @Mapper
 public interface ReservationMapper extends BaseMapper<Reservation> {
     IPage<Reservation> pageReserve(IPage<Reservation> page, @Param(Constants.WRAPPER) Wrapper<Reservation> wrapper);
-    Reservation selectByUserIdAndGoodsId(@Param("userId") Long userId, @Param("goodsId") Integer goodsId);
+    Reservation selectByUserIdAndGoodsId(@Param("userId") Integer userId, @Param("goodsId") Integer goodsId);
 }
