+++ marketback/marketgod/src/main/java/com/th/marketgod/service/ReservationService.java
@@ -18,7 +18,7 @@ public interface ReservationService extends IService<Reservation> {
 
 
     IPage<Reservation> pageReserve(IPage<Reservation> page, Wrapper<Reservation> wrapper);
-    Result reserveGoods(Long userId, Integer goodsId, Integer reserveNum);
-    Result cancelReservation(Long userId, Integer goodsId);
-    boolean checkCommentPermission(Long userId, Integer goodsId);
+    Result reserveGoods(Integer userId, Integer goodsId, Integer reserveNum);
+    Result cancelReservation(Integer userId, Integer goodsId);
+    boolean checkCommentPermission(Integer userId, Integer goodsId);
 }
